requirement.txt
asgiref==3.9.1
Django==5.2.6
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
opencv-python==4.12.0.88
face-recognition==1.3.0
numpy==2.2.6
Pillow==10.4.0
# Alternative: Use deepface for face recognition (easier to install)
# deepface==0.0.79
